requirements.txt
supabase
twilio
python-dotenv
Flask
flask_cors
websockets
flask-sock
openai